simple_requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.50.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
openpyxl>=3.1.0